add TestBase class
